Use lodash isEmpty to check simhash back-end response
Back-end may return `{}` or `[]` when it doesn't have data.
Its better to use lodash `_.isEmpty(value)` which checks if value is
an empty object, collection, map, or set. Instead of checkig for a list
or a dict explicitly. This way, we are more future-proof.

https://lodash.com/docs/4.17.11#isEmpty

Include `lodash` in package.json. Update build and increment pkg
version.