false-positive-diff-util optimisations
`isNotAResource` must be called only once for each `del` and `ins` item.
Previously it was called more times for `ins` elements.

Stop breaking the xpath in array in `removeDiffXPATH`.
Since the `/del[` or `/ins[` pattern is always in the end of the xpath
query. We can just find their location and get a substring of the xpath
query up to that point.