Narrower ymd-timestamp-header select controls
The `<select>` first values: "Months and available captures"  "Years and
available captures" are too wide and make the UI span in 2 lines.
That doesn't look good.  (I agree that it conveys much more info but it
doesn't look that good).
We make the options "Month" and "Year" and everything fits into
one line.
The previous options are set to `title=""` attributes of the select elements.

Bump version to 0.2.18 and update build files.