Fix CDX server endpoint
We used to append `search` on CDX server endpoint. This is not
necessary. We define the CDX in conf, there is no need to add something
else in the code, it causes trouble. For instance: when `cdxServer` is
`/web/timemap/cdx`, the final CDX query would be like
`/web/timemap/cdx/search` which doesn't exist.